requirements.txt
asgiref==3.5.0
bleach==4.1.0
certifi==2021.10.8
charset-normalizer==2.0.11
Django==4.0.2
django-cors-headers==3.11.0
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
Faker==8.13.2
idna==3.3
packaging==21.3
PyJWT==2.1.0
pyparsing==3.0.7
python-dateutil==2.8.2
python-decouple==3.4
pytz==2021.3
requests==2.27.1
six==1.16.0
sqlparse==0.4.2
text-unidecode==1.3
urllib3==1.26.8
webencodings==0.5.1
